backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pgvector>=0.2.4

# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Auth
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# AI
openai>=1.10.0

# Storage (Cloudflare R2)
boto3>=1.34.25

# Integrations
composio-core>=0.7.0

# Monitoring
sentry-sdk[fastapi]>=2.0.0

# Rate Limiting
slowapi>=0.1.9

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
